refactor: extract indicator mappers into testable mapIndicators module
- Extract mapIndicatorA, mapIndicatorB, mapIndicatorF, mapCseOpinions
  from buildExportRows.ts into mapIndicators.ts (pure functions, no DB)
- Add 9 unit tests for all mappers (A, B, F, CSE) covering happy path,
  empty input, and wrong step filtering
- buildExportRows.ts reduced to DB queries + assembly (~200 lines)
